[#2828] Allow decimals for the number of days in the finished configuration.

+++ app/models/kanban.rb
@@ -82,7 +82,7 @@ class Kanban
     days = @settings['panes']['finished']['limit'] || 7
     issues = Issue.visible.all(:include => :assigned_to,
                                :order => "#{Issue.table_name}.updated_on DESC",
-                               :conditions => ["#{Issue.table_name}.status_id = ? AND #{Issue.table_name}.updated_on > ?", @settings['panes']['finished']['status'], days.to_i.days.ago])
+                               :conditions => ["#{Issue.table_name}.status_id = ? AND #{Issue.table_name}.updated_on > ?", @settings['panes']['finished']['status'], days.to_f.days.ago])
 
     return issues.group_by(&:assigned_to)
   end
